requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
